// src/app/type.d.ts
export type MaterialConfigId = string;
export type FinishGroupId = string;
export type MaterialGroupId = string;
export type MaterialId = string;
export type QuoteId = string;
export type VendorId = string;
export type ConfigId = string;
export type ShippingId = string;
export type ModelId = string;
// ...
export type PrintingService = {
  [vendorId: string]: {
    materialId: MaterialId;
    finishId: FinishGroupId;
    printingMethodShort: string;
    printingMethod: string;
    materialName: string;
    productionTimeFast: number;
    productionTimeSlow: number;
  };
};

export type MaterialConfig = {
  id: MaterialConfigId;
  name: string;
  color: string;
  colorCode: string;
  colorImage: string;
  printingService: PrintingService;
  finishGroupId: FinishGroupId;
  materialId: MaterialId;
  materialGroupId: MaterialGroupId;
};
// ...
export type Quote = {
  quoteId: QuoteId;
  vendorId: VendorId;
  modelId: ModelId;
  materialConfigId: MaterialConfigId;
  price: number;
  grossPrice: number;
  quantity: number;
  currency: string;
  isPrintable: boolean;
};

export type MultiModelQuote = {
  vendorId: VendorId;
  materialConfigId: MaterialConfigId;
  price: number;
  grossPrice: number;
  currency: string;
  isPrintable: boolean;
  quotes: Array<Quote>;
};

export type Shipping = {
  shippingId: ShippingId;
  vendorId: VendorId;
  name: string;
  deliveryTime: string;
  price: number;
  grossPrice: number;
  currency: string;
};

export type Offer = {
  multiModelQuote: MultiModelQuote;
  shipping: Shipping;
  totalGrossPrice: number;
};